execution/requirements.txt
# ── Core Web Framework ────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── AI ────────────────────────────────────────────────────────
openai>=1.30.0

# ── Environment ───────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Scraping ──────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Data Validation ───────────────────────────────────────────
pydantic>=2.0.0
